bill-validator-backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.6
aiohttp==3.10.10
python-dotenv==1.1.1
pydantic==2.11.7
Pillow==10.4.0
requests==2.32.5